+++ src/main/java/walkingkooka/spreadsheet/dominokit/history/SpreadsheetMetadataSelectHistoryToken.java
@@ -22,6 +22,7 @@ import walkingkooka.spreadsheet.SpreadsheetId;
 import walkingkooka.spreadsheet.SpreadsheetName;
 import walkingkooka.spreadsheet.dominokit.AppContext;
 import walkingkooka.spreadsheet.format.pattern.SpreadsheetPatternKind;
+import walkingkooka.spreadsheet.meta.SpreadsheetMetadataPropertyName;
 import walkingkooka.tree.text.TextStylePropertyName;
 
 import java.util.Optional;
@@ -61,6 +62,15 @@ public final class SpreadsheetMetadataSelectHistoryToken extends SpreadsheetMeta
         return with(
                 id,
                 name
+        );
+    }
+
+    @Override
+    HistoryToken setMetadataPropertyName0(final SpreadsheetMetadataPropertyName<?> propertyName) {
+        return metadataPropertySelect(
+                this.id(),
+                this.name(),
+                propertyName
         );
     }
